ENH: Generate pure plugs mask and its incorporation in BABC
An ITK filter is added to evaluate a sample set to check whether they
belong to a pure plug or not within the defined threshold.
A Mahalanobis-weighted Euclidean distance is used as the metric.
This membership filter is called in a program that collects samples as
normalized intensity values (between 0 and 1) of the input multimodal images.
Note that it is assumed that input multimodal scans are aligned in
physical space.

Usage:
$ bin/GeneratePurePlugMask \
  --inputImageModalities t1.nii.gz,t2.nii.gz,IDWI.nrrd \
  --threshold 0.2 \
  --numberOfSubSamples 2,2,2 \
  --outputMaskFile binary_pure_mask.nrrd

GeneratePurePlugMask.h is called to compute pure plugs, and this file
calls "itkIntegrityMetricMembershipFunction" filter to decide each plug
is pure or not.

Also, many changes in BRAINSABC are refactoring BABC such that it can
perform in physical space rather than voxel space.